@@ -1,16 +1,15 @@
 name: swc_flutter
 description: flutter developement simplifier
-version: 1.4.1
+version: 1.5.0-nullsafe
 homepage: https://github.com/SBNTT/swc_flutter
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-
-  provider: ^4.1.2
+  provider: ^5.0.0
 
 dev_dependencies:
   flutter_test:
