@@ -5,7 +5,7 @@ import 'package:swc_flutter/src/swc_widget.dart';
 class SwcWidgetState<C extends SwcController<SwcWidget>> {
 
   bool initialized = false;
-  C controller;
-  List<SingleChildWidget> providers;
+  late C controller;
+  late List<SingleChildWidget> providers;
   
 }
