@@ -5,9 +5,9 @@ import 'package:swc_flutter/swc_flutter.dart';
 import 'common.dart';
 
 void main() {
-  group("Test SwcStatelessWidget", () {
-    testWidgets("Wrapping a hardcoded Text Widget", (tester) async {
-      final content = "Hi!";
+  group('Test SwcStatelessWidget', () {
+    testWidgets('Wrapping a hardcoded Text Widget', (tester) async {
+      final content = 'Hi!';
       await tester.pumpWidget(MaterialApp(
         home: _SimpleTestWidget(content: content),
       ));
@@ -15,22 +15,22 @@ void main() {
       expect(find.text(content), findsOneWidget);
     });
 
-    testWidgets("Counter widget", (tester) async {
+    testWidgets('Counter widget', (tester) async {
       await tester.pumpWidget(MaterialApp(
         home: _CounterTestWidget(),
       ));
 
-      expect(find.text("0"), findsOneWidget);
-      expect(find.text("1"), findsNothing);
+      expect(find.text('0'), findsOneWidget);
+      expect(find.text('1'), findsNothing);
 
       await tester.tap(find.byIcon(Icons.add));
       await tester.pump();
 
-      expect(find.text("0"), findsNothing);
-      expect(find.text("1"), findsOneWidget);
+      expect(find.text('0'), findsNothing);
+      expect(find.text('1'), findsOneWidget);
     });
 
-    testWidgets("Access to widget within SwcController", (tester) async {
+    testWidgets('Access to widget within SwcController', (tester) async {
       await tester.pumpWidget(MaterialApp(
         home: _CounterTestWidget(),
       ));
@@ -44,17 +44,20 @@ void main() {
   });
 }
 
-class _SimpleTestWidget extends SwcStatelessWidget {
+class _SimpleTestWidget extends SwcStatelessWidget<EmptyController> {
 
-  final String content;
+  final String? content;
 
   _SimpleTestWidget({this.content});
 
-  getController() => null;
+  @override
+  getController() => EmptyController();
 
+  @override
   getProviders() => [];
 
-  build(_) => Text(content);
+  @override
+  build(_) => Text(content!);
 
 }
 
@@ -62,20 +65,23 @@ class _CounterTestWidget extends SwcStatelessWidget<CounterController> {
 
   final _counterState = CounterState();
 
+  @override
   getController() => CounterController();
 
+  @override
   getProviders() => [
     ChangeNotifierProvider<CounterState>.value(value: _counterState),
   ];
 
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: Consumer<CounterState>(builder: (context, state, _) {
         return Text(state.get().toString());
       }),
       floatingActionButton: FloatingActionButton(
-        child: const Icon(Icons.add),
         onPressed: () => controller.onAddButtonClick(context),
+        child: const Icon(Icons.add),
       ),
     );
   }
