@@ -1,6 +1,6 @@
 library swc_flutter;
 
-export 'src/swc_stateless_widget.dart';
-export 'src/swc_stateful_widget.dart';
-export 'src/swc_controller.dart';
 export 'src/provider_helper.dart';
+export 'src/swc_controller.dart';
+export 'src/swc_stateful_widget.dart';
+export 'src/swc_stateless_widget.dart';
